jobs/sync_wordpress_posts/requirements.txt
google-auth==2.41.1
google-cloud-bigquery==3.38.0
requests==2.32.5
